backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.18
azure-ai-documentintelligence==1.0.0b2
python-dotenv==1.0.1
pydantic==2.6.1
pytest==8.0.0
httpx==0.26.0
